Add authorization system.
Getting logs for a channel now requires the 'logging' permission in
that channel. This is to prevent random folk from getting histories of
private channels.

Pulled help system out into a plugin.

There's now the notion of 'core' plugins, which load in order. This
way, you can have (acyclic!) dependencies among core plugins, and
normal plugins can depend on everything in core without fear.

Fixed a bug or two in 00-brain.js; it had nobody using it before, so
there was some dust to brush off.