Log what the bot says.
The initial implementation of this just listened to the 'selfMessage'
event and logged it. However, if a plugin listened to the 'message'
event and talked back to the user with the say() method, then the
reply would appear in the log before the request. This is because
'message' is emitted before 'message#', which is what the logging
plugin was using.

The fix is twofold: one, load logging first (so now plugin loads are
ordered and it's named 00-logging.js, thus it loads first) and have
logging listen to the 'message' event, not 'message#'.